chore: fix lodash imports, move hooks
Import only lodash modules in use, to reduce size of lodash package
Refactor Hooks.js from config to utils
