+++ nerdlets/find-user/components/session/SessionTable.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { isEqual } from 'lodash'
+import isEqual from 'lodash.isequal'
 import {
   BlockText,
   Spinner,
