+++ nerdlets/mande-nerdlet/components/category-menu/style.scss
@@ -15,9 +15,9 @@
 
 .category-menu-title {
   box-sizing: border-box;
-  padding: 0 16px;
+  padding: .5rem 1rem;
   width: 100%;
-  height: 58px;
+  height: 62px;
   background-color: #edeeee;
 
   font-weight: 600;
@@ -38,7 +38,8 @@
 
 [class*='wnd-Stack'].category-menu-item-content {
   box-sizing: border-box;
-  height: 55px;
+  min-height: 55px;
+  max-height: 80px;
   border-bottom: 1px dotted #e3e4e4;
   justify-content: space-between;
   padding: 0 10px 0 16px;
