+++ nerdlets/video-session/components/timeline/gauge/Gauge.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 
-import _ from 'lodash'
 import { Stack, StackItem } from 'nr1'
 
 import GaugeDataLegendItem from './GaugeDataLegendItem'
