+++ package.json
@@ -2,7 +2,7 @@
   "private": true,
   "name": "nr1-mande",
   "description": "Nerdpack for the Media and Entertainment vertical, built against the internal Video Demo account.",
-  "version": "0.4.12",
+  "version": "0.4.13",
   "scripts": {
     "start": "nr1 nerdpack:serve",
     "test": "exit 0",
@@ -14,7 +14,13 @@
     "graphql": "^14.7.0",
     "graphql-tag": "^2.11.0",
     "hsl-to-hex": "^1.0.0",
-    "lodash": "^4.17.20",
+    "lodash.clonedeep": "^4.5.0",
+    "lodash.flatten": "^4.4.0",
+    "lodash.isequal": "^4.5.0",
+    "lodash.map": "^4.6.0",
+    "lodash.sortby": "^4.7.0",
+    "lodash.startcase": "^4.4.0",
+    "lodash.uniq": "^4.5.0",
     "moment": "^2.27.0",
     "moment-duration-format": "^2.3.2",
     "moment-timezone": "^0.5.31",
